String quoting moved into separate wrapper class
Instead of the ASEDecl.add() method including an extra parameter for quoting
strings, this functionality is now in a separate Quoted class which
automatically adds quotes to any string it is constructed with. As well as
potentially being useful elsewhere, this makes the method invocations more
self-explanatory: "decl.add('KEY', Quoted('value'))"

Also switched some uses of __repr__ to __str__ instead, in accordance with
Python conventions (__repr__ is intended for returning a "formal"
representation of the object which could be used to construct a new instance
with eval()).